+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [9.90.4] - 2019-10-25
+
+### Added
+
+- **NumericStepper** `readOnly` prop.
+
 ## [9.90.3] - 2019-10-25
 
 ### Fixed